Restructure commands into single files
