+++ functions/commands/getcamerastream.js
@@ -1,0 +1,28 @@
+const DefaultCommand = require('./default.js');
+
+class GetCameraStream extends DefaultCommand {
+  static get type() {
+    return 'action.devices.commands.GetCameraStream';
+  }
+
+  static validateParams(params) {
+    return ('StreamToChromecast' in params) && typeof params.StreamToChromecast === 'boolean' &&
+      ('SupportedStreamProtocols' in params) && typeof params.SupportedStreamProtocols === 'object';
+  }
+
+  static get requiresItem() {
+    return true;
+  }
+
+  static convertParamsToValue() {
+    return null;
+  }
+
+  static getResponseStates(params, item) {
+    return {
+      cameraStreamAccessUrl: item.state
+    };
+  }
+}
+
+module.exports = GetCameraStream;
