+++ functions/commands/setvolume.js
@@ -1,0 +1,24 @@
+const DefaultCommand = require('./default.js');
+
+class SetVolume extends DefaultCommand {
+  static get type() {
+    return 'action.devices.commands.setVolume';
+  }
+
+  static validateParams(params) {
+    return ('volumeLevel' in params) && typeof params.volumeLevel === 'number';
+  }
+
+  static convertParamsToValue(params) {
+    return params.volumeLevel.toString();
+  }
+
+  static getResponseStates(params) {
+    return {
+      currentVolume: params.volumeLevel,
+      isMuted: params.volumeLevel === 0
+    };
+  }
+}
+
+module.exports = SetVolume;
