+++ functions/commands/thermostattemperaturesetpoint.js
@@ -1,0 +1,40 @@
+const DefaultCommand = require('./default.js');
+const Thermostat = require('../devices/thermostat.js');
+
+class ThermostatTemperatureSetpoint extends DefaultCommand {
+  static get type() {
+    return 'action.devices.commands.ThermostatTemperatureSetpoint';
+  }
+
+  static validateParams(params) {
+    return ('thermostatTemperatureSetpoint' in params) && typeof params.thermostatTemperatureSetpoint === 'number';
+  }
+
+  static get requiresItem() {
+    return true;
+  }
+
+  static getItemName(item) {
+    const members = Thermostat.getMembers(item);
+    if ('thermostatTemperatureSetpoint' in members) {
+      return members.thermostatTemperatureSetpoint.name;
+    }
+    throw { statusCode: 400 };
+  }
+
+  static convertParamsToValue(params, item) {
+    let value = params.thermostatTemperatureSetpoint;
+    if (Thermostat.usesFahrenheit(item)) {
+      value = Thermostat.convertToFahrenheit(value);
+    }
+    return value.toString();
+  }
+
+  static getResponseStates(params, item) {
+    const states = Thermostat.getState(item);
+    states.thermostatTemperatureSetpoint = params.thermostatTemperatureSetpoint;
+    return states;
+  }
+}
+
+module.exports = ThermostatTemperatureSetpoint;
